add the codex app architecture
